// types/database.types.ts
export type Json =
  | string
  | number
  | boolean
  | null
  | { [key: string]: Json | undefined }
  | Json[]

export type Database = {
  public: {
    Tables: {
// ...
      resturents: {
        Row: {
          activities: string[] | null
          address: string | null
          age_restriction: number | null
          created_at: string
          email: string | null
          featured_image: string | null
          id: string
          images: Json | null
          is_facilitated: boolean | null
          location: unknown
          logo: string | null
          name: string | null
          opening_hours: Json | null
          price: string | null
          servings: string[] | null
          website: string | null
        }
        Insert: {
          activities?: string[] | null
          address?: string | null
          age_restriction?: number | null
          created_at?: string
          email?: string | null
          featured_image?: string | null
          id?: string
          images?: Json | null
          is_facilitated?: boolean | null
          location: unknown
          logo?: string | null
          name?: string | null
          opening_hours?: Json | null
          price?: string | null
          servings?: string[] | null
          website?: string | null
        }
        Update: {
          activities?: string[] | null
          address?: string | null
          age_restriction?: number | null
          created_at?: string
          email?: string | null
          featured_image?: string | null
          id?: string
          images?: Json | null
          is_facilitated?: boolean | null
          location?: unknown
          logo?: string | null
          name?: string | null
          opening_hours?: Json | null
          price?: string | null
          servings?: string[] | null
          website?: string | null
        }
        Relationships: []
      }
      serving: {
        Row: {
          created_at: string
          iconUrl: string | null
          id: number
          name: string | null
        }
        Insert: {
          created_at?: string
          iconUrl?: string | null
          id?: number
          name?: string | null
        }
        Update: {
          created_at?: string
          iconUrl?: string | null
          id?: number
          name?: string | null
        }
        Relationships: []
      }
// ...
    }
// ...
  }
}